Update patch set 5
Patch Set 5:

I think we should use a single chart and use a value setting to determine if the go or python adapter and its options are used. I think this would better facilitate an upgrade to a single adapter at a later point. The changes to the existing chart to support this should be minimal.

Patch-set: 5
Reviewer: Gerrit User 1000003 <1000003@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0