Update patch set 2
Patch Set 2:

Attempting to understand the "open onu" adapter. I thought the idea was that the "open onu" adapter would work for any device on which the "open" agent side of this was running. In other words the adapter was generic and the specifics resided on the side of the device? Is this not the case? If we are to have a brcm_open_onu and an X,Y, or Z open_onu adapter it seems that it is just an adapter. What am I missing?

Patch-set: 2
Label: Verified=0