Update patch set 2
Patch Set 2:

> > system-wide configuration should, i think be stored in the KV
 > > store, especially configuration that can be modified at runtime,
 > > such as log levels. for example, if i set model logging at 3 and
 > > start voltha, then set it to 5 via the API, and then core
 > restarts
 > > the logging is not back at 3. I don't think a log file is the
 > best
 > > approach.
 > 
 > We already have a system-wide log_level parameter. This patch
 > brings flexibility at packet level. It can be useful until the
 > better approach. Otherwise people need to compile the code if they
 > want to change a package's log level. It would not hide the need of
 > dynamic log level mechanism - we still need it.

The existing system-wide logging level was sort of the compromise (i.e. i know we really should to X, but lets just to this until X is implemented.) So I am hesitant to continue with another compromise. if this is a useful capability then we need to implement it as we think it should be in the long term.

Patch-set: 2