Update patch set 1
Patch Set 1:

Could you include the Jira issue (VOL-####) for this code as the first line of the commit message?

Because of how the helm chart publisher works, any change to a helm chart will require increasing the version in Chart.yaml - please increase the minor version in this patchset as it adds a feature.

Patch-set: 1
Label: Verified=0