Update patch set 2
Patch Set 2:

the hardware test is unstable due to an issue with one of the ONUs on the Tuscon pod. Merging this patch now as this was manually verified to be working on one of the ONF demo pods.

Patch-set: 2
Reviewer: Gerrit User 1000127 <1000127@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>